+++ package.json
@@ -16,7 +16,7 @@
     ]
   },
   "dependencies": {
-    "@ybot1122/toby-ui": "0.0.14-alpha",
+    "@ybot1122/toby-ui": "0.0.15-alpha",
     "next": "15.1.4",
     "react": "19.0.0",
     "react-dom": "19.0.0"
